Ran repo through clang-Tidy
1)removed useless const qualifiers from method declarations/return types
2)made some methods static that don't need a class object to work
3)General typos

+++ src/player.cpp
@@ -13,7 +13,7 @@ Player::Player(const int mark, const std::string_view name)
 //Move Logic
 int Player::moves() const
 {
-    return m_moves.size();
+    return static_cast<int>(m_moves.size());
 }
 
 bool Player::isOccupying(const int move) const
